+++ src/docs/pages/component-docs/components/drawer/drawer-nav-group/drawer-nav-group-doc.component.ts
@@ -1,14 +1,15 @@
-import { Component } from '@angular/core';
-import { MarkdownSplitService } from '../../../../../services/markdown-split/markdown-split.service';
-import { MarkdownService } from 'ngx-markdown';
-import { BASIC } from './examples/basic.component';
-import { MULTIPLE_GROUPS } from './examples/multiple-groups.component';
-import { SPACER } from './examples/with-spacer.component';
+import {Component} from '@angular/core';
+import {MarkdownSplitService} from '../../../../../services/markdown-split/markdown-split.service';
+import {MarkdownService} from 'ngx-markdown';
+import {BASIC} from './examples/basic.component';
+import {MULTIPLE_GROUPS} from './examples/multiple-groups.component';
+import {SPACER} from './examples/with-spacer.component';
+import {NavGroupPlaygroundKnobs} from "./examples/playground.component";
 
 @Component({
     selector: 'app-drawer-nav-group-doc',
     template: `
-        <app-component-doc-scaffold [md]="md">
+        <app-component-doc-scaffold [md]="md" [knobs]="knobs">
             <div examples class="app-example">
                 <div class="example-section">
                     <div class="example-heading">Basic Drawer Nav Group</div>
@@ -53,9 +54,13 @@ import { SPACER } from './examples/with-spacer.component';
                     </div>
                 </div>
             </div>
-            <div playground></div>
-            <div code></div>
-            <div knobs></div>
+
+            <app-nav-group-playground
+                    playground
+                    [inputs]="knobs"
+                    (codeChange)="generatedCode = $event"
+            ></app-nav-group-playground>
+            <app-example-code code [snippet]="generatedCode"></app-example-code>
         </app-component-doc-scaffold>
     `,
     styleUrls: ['./drawer-nav-group-doc.component.scss'],
@@ -65,6 +70,22 @@ export class DrawerNavGroupDocComponent {
     BASIC = BASIC;
     MULTIPLE_GROUPS = MULTIPLE_GROUPS;
     SPACER = SPACER;
+    generatedCode: string;
+
+    /* Default playground knobs */
+    knobs: NavGroupPlaygroundKnobs = {
+        title: {
+            value: 'Group 1',
+            type: 'string',
+            hint: 'Title that appears in above nav items',
+        },
+        divider: {
+            value: true,
+            componentDefault: true,
+            type: 'boolean',
+            hint: 'Show divider under nav group',
+        },
+    };
 
     constructor(
         private readonly _splitService: MarkdownSplitService,