+++ src/docs/pages/component-docs/components/drawer/drawer-nav-group/drawer-nav-group-doc.module.ts
@@ -5,9 +5,10 @@ import { DrawerNavGroupDocComponent } from './drawer-nav-group-doc.component';
 import { BasicExampleComponent } from './examples/basic.component';
 import { MultipleGroupsComponent } from './examples/multiple-groups.component';
 import { WithSpacerComponent } from './examples/with-spacer.component';
+import {PlaygroundComponent} from "./examples/playground.component";
 
 @NgModule({
-    declarations: [DrawerNavGroupDocComponent, BasicExampleComponent, MultipleGroupsComponent, WithSpacerComponent],
+    declarations: [DrawerNavGroupDocComponent, BasicExampleComponent, MultipleGroupsComponent, WithSpacerComponent, PlaygroundComponent],
     imports: [DrawerModule, SharedCompDocsModule],
     exports: [DrawerNavGroupDocComponent],
 })